test(studio-api): split server tests by route domain
