+++ packages/studio/src/api/server.daemon-doctor.test.ts
@@ -1,0 +1,174 @@
+import { describe, expect, it, vi } from "vitest";
+import {
+  root,
+  schedulerStartMock,
+  runRadarMock,
+  createLLMClientMock,
+  chatCompletionMock,
+  loadProjectConfigMock,
+  pipelineConfigs,
+  cloneProjectConfig
+} from "./__tests__/server-test-utils";
+
+describe("daemon doctor", () => {
+  it("returns from /api/daemon/start before the first write cycle finishes", async () => {
+    let resolveStart: (() => void) | undefined;
+    schedulerStartMock.mockImplementation(
+      () =>
+        new Promise<void>((resolve) => {
+          resolveStart = resolve;
+        }),
+    );
+
+    const { createStudioServer } = await import("./server.js");
+    const app = createStudioServer(cloneProjectConfig() as never, root);
+
+    const responseOrTimeout = await Promise.race([
+      app.request("http://localhost/api/v1/daemon/start", { method: "POST" }),
+      new Promise<"timeout">((resolve) => setTimeout(() => resolve("timeout"), 30)),
+    ]);
+
+    expect(responseOrTimeout).not.toBe("timeout");
+
+    const response = responseOrTimeout as Response;
+    expect(response.status).toBe(200);
+    await expect(response.json()).resolves.toMatchObject({ ok: true, running: true });
+
+    const status = await app.request("http://localhost/api/v1/daemon");
+    await expect(status.json()).resolves.toEqual({ running: true });
+
+    resolveStart?.();
+  });
+
+  it("reloads latest llm config for doctor checks without restarting the studio server", async () => {
+    const startupConfig = {
+      ...cloneProjectConfig(),
+      llm: {
+        ...cloneProjectConfig().llm,
+        model: "stale-model",
+        baseUrl: "https://stale.example.com/v1",
+      },
+    };
+
+    const freshConfig = {
+      ...cloneProjectConfig(),
+      llm: {
+        ...cloneProjectConfig().llm,
+        model: "fresh-model",
+        baseUrl: "https://fresh.example.com/v1",
+      },
+    };
+    loadProjectConfigMock.mockResolvedValue(freshConfig);
+
+    // Stub /models so probe doesn't hit the real OpenAI endpoint and short-circuit on 401.
+    const fetchMock = vi.fn().mockResolvedValue({
+      ok: false,
+      status: 404,
+      text: async () => "Not Found",
+    });
+    vi.stubGlobal("fetch", fetchMock as typeof fetch);
+
+    const { createStudioServer } = await import("./server.js");
+    const app = createStudioServer(startupConfig as never, root);
+
+    const response = await app.request("http://localhost/api/v1/doctor");
+
+    expect(response.status).toBe(200);
+    expect(createLLMClientMock).toHaveBeenCalledWith(expect.objectContaining({
+      model: "fresh-model",
+      baseUrl: "https://fresh.example.com/v1",
+    }));
+    expect(chatCompletionMock).toHaveBeenCalledWith(
+      expect.anything(),
+      "fresh-model",
+      expect.any(Array),
+      expect.objectContaining({ maxTokens: expect.any(Number) }),
+    );
+  });
+
+  it("auto-falls back to a non-stream probe in doctor checks when the first transport returns empty", async () => {
+    const freshConfig = {
+      ...cloneProjectConfig(),
+      llm: {
+        ...cloneProjectConfig().llm,
+        model: "claude-sonnet-4-6",
+        baseUrl: "https://timesniper.club",
+        stream: true,
+        apiFormat: "chat",
+      },
+    };
+    loadProjectConfigMock.mockResolvedValue(freshConfig);
+    // Stub /models so probe doesn't hit the real OpenAI endpoint and short-circuit on 401.
+    const fetchMock = vi.fn().mockResolvedValue({
+      ok: false,
+      status: 404,
+      text: async () => "Not Found",
+    });
+    vi.stubGlobal("fetch", fetchMock as typeof fetch);
+    createLLMClientMock.mockImplementation(((cfg: unknown) => cfg) as any);
+    chatCompletionMock.mockImplementation(async (client: any) => {
+      if (client.stream === false) {
+        return {
+          content: "pong",
+          usage: { promptTokens: 1, completionTokens: 1, totalTokens: 2 },
+        };
+      }
+      throw new Error("LLM returned empty response from stream");
+    });
+
+    const { createStudioServer } = await import("./server.js");
+    const app = createStudioServer(freshConfig as never, root);
+
+    const response = await app.request("http://localhost/api/v1/doctor");
+    expect(response.status).toBe(200);
+    await expect(response.json()).resolves.toMatchObject({
+      llmConnected: true,
+    });
+    expect(createLLMClientMock).toHaveBeenCalledWith(expect.objectContaining({
+      stream: true,
+      apiFormat: "chat",
+    }));
+    expect(createLLMClientMock).toHaveBeenCalledWith(expect.objectContaining({
+      stream: false,
+      apiFormat: "chat",
+    }));
+  });
+
+  it("reloads latest llm config for radar scans without restarting the studio server", async () => {
+    const startupConfig = {
+      ...cloneProjectConfig(),
+      llm: {
+        ...cloneProjectConfig().llm,
+        model: "stale-model",
+        baseUrl: "https://stale.example.com/v1",
+      },
+    };
+
+    const freshConfig = {
+      ...cloneProjectConfig(),
+      llm: {
+        ...cloneProjectConfig().llm,
+        model: "fresh-model",
+        baseUrl: "https://fresh.example.com/v1",
+      },
+    };
+    loadProjectConfigMock.mockResolvedValue(freshConfig);
+
+    const { createStudioServer } = await import("./server.js");
+    const app = createStudioServer(startupConfig as never, root);
+
+    const response = await app.request("http://localhost/api/v1/radar/scan", {
+      method: "POST",
+    });
+
+    expect(response.status).toBe(200);
+    expect(runRadarMock).toHaveBeenCalledTimes(1);
+    expect(pipelineConfigs.at(-1)).toMatchObject({
+      model: "fresh-model",
+      defaultLLMConfig: expect.objectContaining({
+        model: "fresh-model",
+        baseUrl: "https://fresh.example.com/v1",
+      }),
+    });
+  });
+});
