+++ packages/studio/src/api/server.test.ts
@@ -1,90 +1,16 @@
 import { describe, expect, it, vi } from "vitest";
-import { access, mkdir, readFile, writeFile } from "node:fs/promises";
-import { join } from "node:path";
-import { tmpdir } from "node:os";
 import {
   root,
-  schedulerStartMock,
-  initBookMock,
-  runRadarMock,
-  reviseDraftMock,
-  resyncChapterArtifactsMock,
-  writeNextChapterMock,
-  rollbackToChapterMock,
-  saveChapterIndexMock,
-  loadChapterIndexMock,
-  loadBookConfigMock,
-  createLLMClientMock,
-  chatCompletionMock,
-  loadProjectConfigMock,
-  pipelineConfigs,
-  processProjectInteractionInputMock,
-  processProjectInteractionRequestMock,
-  createInteractionToolsFromDepsMock,
-  loadProjectSessionMock,
-  resolveSessionActiveBookMock,
-  runAgentSessionMock,
-  createAndPersistBookSessionMock,
-  loadBookSessionMock,
-  persistBookSessionMock,
-  appendBookSessionMessageMock,
-  appendManualSessionMessagesMock,
-  renameBookSessionMock,
-  deleteBookSessionMock,
-  migrateBookSessionMock,
-  resolveServiceModelMock,
-  loadSecretsMock,
-  saveSecretsMock,
-  getServiceApiKeyMock,
-  SERVICE_PRESETS_MOCK,
-  resolveServicePresetMock,
-  resolveServiceProviderFamilyMock,
-  resolveServiceModelsBaseUrlMock,
-  listModelsForServiceMock,
-  endpointIdsByGroup,
-  endpointMocks,
-  getAllEndpointsMock,
-  probeModelsFromUpstreamMock,
-  projectConfig,
   cloneProjectConfig
 } from "./__tests__/server-test-utils";
 
-describe("createStudioServer daemon lifecycle", () => {
+describe("createStudioServer safety guards", () => {
   it("uses the real core bookId validator in the Studio safety mock", async () => {
     const { isSafeBookId } = await import("@actalk/inkos-core");
 
     expect(vi.isMockFunction(isSafeBookId)).toBe(false);
     expect(isSafeBookId("demo-book")).toBe(true);
     expect(isSafeBookId("demo/book")).toBe(false);
-  });
-
-  it("returns from /api/daemon/start before the first write cycle finishes", async () => {
-    let resolveStart: (() => void) | undefined;
-    schedulerStartMock.mockImplementation(
-      () =>
-        new Promise<void>((resolve) => {
-          resolveStart = resolve;
-        }),
-    );
-
-    const { createStudioServer } = await import("./server.js");
-    const app = createStudioServer(cloneProjectConfig() as never, root);
-
-    const responseOrTimeout = await Promise.race([
-      app.request("http://localhost/api/v1/daemon/start", { method: "POST" }),
-      new Promise<"timeout">((resolve) => setTimeout(() => resolve("timeout"), 30)),
-    ]);
-
-    expect(responseOrTimeout).not.toBe("timeout");
-
-    const response = responseOrTimeout as Response;
-    expect(response.status).toBe(200);
-    await expect(response.json()).resolves.toMatchObject({ ok: true, running: true });
-
-    const status = await app.request("http://localhost/api/v1/daemon");
-    await expect(status.json()).resolves.toEqual({ running: true });
-
-    resolveStart?.();
   });
 
   it("rejects book routes with path traversal ids", async () => {
@@ -101,2201 +27,6 @@ describe("createStudioServer daemon lifecycle", () => {
         code: "INVALID_BOOK_ID",
         message: 'Invalid book ID: "../etc/passwd"',
       },
-    });
-  });
-
-  it("allows reading and updating fixed control truth files", async () => {
-    const bookDir = join(root, "books", "demo-book");
-    const storyDir = join(bookDir, "story");
-    await mkdir(storyDir, { recursive: true });
-    await Promise.all([
-      writeFile(join(storyDir, "author_intent.md"), "# Author Intent\n\nStay cold.\n", "utf-8"),
-      writeFile(join(storyDir, "current_focus.md"), "# Current Focus\n\nReturn to the old case.\n", "utf-8"),
-    ]);
-
-    const { createStudioServer } = await import("./server.js");
-    const app = createStudioServer(cloneProjectConfig() as never, root);
-
-    const readAuthorIntent = await app.request("http://localhost/api/v1/books/demo-book/truth/author_intent.md");
-    expect(readAuthorIntent.status).toBe(200);
-    await expect(readAuthorIntent.json()).resolves.toMatchObject({
-      file: "author_intent.md",
-      content: "# Author Intent\n\nStay cold.\n",
-    });
-
-    const updateCurrentFocus = await app.request("http://localhost/api/v1/books/demo-book/truth/current_focus.md", {
-      method: "PUT",
-      headers: { "Content-Type": "application/json" },
-      body: JSON.stringify({ content: "# Current Focus\n\nPull focus back to the harbor trail.\n" }),
-    });
-    expect(updateCurrentFocus.status).toBe(200);
-
-    await expect(readFile(join(storyDir, "current_focus.md"), "utf-8")).resolves.toBe(
-      "# Current Focus\n\nPull focus back to the harbor trail.\n",
-    );
-  });
-
-  it("reflects project edits immediately without restarting the studio server", async () => {
-    const { createStudioServer } = await import("./server.js");
-    const app = createStudioServer(cloneProjectConfig() as never, root);
-
-    const save = await app.request("http://localhost/api/v1/project", {
-      method: "PUT",
-      headers: { "Content-Type": "application/json" },
-      body: JSON.stringify({
-        language: "en",
-        temperature: 0.2,
-        stream: true,
-      }),
-    });
-
-    expect(save.status).toBe(200);
-
-    const project = await app.request("http://localhost/api/v1/project");
-    await expect(project.json()).resolves.toMatchObject({
-      language: "en",
-      temperature: 0.2,
-      stream: true,
-    });
-  });
-
-  it("reloads latest llm config for doctor checks without restarting the studio server", async () => {
-    const startupConfig = {
-      ...cloneProjectConfig(),
-      llm: {
-        ...cloneProjectConfig().llm,
-        model: "stale-model",
-        baseUrl: "https://stale.example.com/v1",
-      },
-    };
-
-    const freshConfig = {
-      ...cloneProjectConfig(),
-      llm: {
-        ...cloneProjectConfig().llm,
-        model: "fresh-model",
-        baseUrl: "https://fresh.example.com/v1",
-      },
-    };
-    loadProjectConfigMock.mockResolvedValue(freshConfig);
-
-    // Stub /models so probe doesn't hit the real OpenAI endpoint and short-circuit on 401.
-    const fetchMock = vi.fn().mockResolvedValue({
-      ok: false,
-      status: 404,
-      text: async () => "Not Found",
-    });
-    vi.stubGlobal("fetch", fetchMock as typeof fetch);
-
-    const { createStudioServer } = await import("./server.js");
-    const app = createStudioServer(startupConfig as never, root);
-
-    const response = await app.request("http://localhost/api/v1/doctor");
-
-    expect(response.status).toBe(200);
-    expect(createLLMClientMock).toHaveBeenCalledWith(expect.objectContaining({
-      model: "fresh-model",
-      baseUrl: "https://fresh.example.com/v1",
-    }));
-    expect(chatCompletionMock).toHaveBeenCalledWith(
-      expect.anything(),
-      "fresh-model",
-      expect.any(Array),
-      expect.objectContaining({ maxTokens: expect.any(Number) }),
-    );
-  });
-
-  it("auto-falls back to a non-stream probe in doctor checks when the first transport returns empty", async () => {
-    const freshConfig = {
-      ...cloneProjectConfig(),
-      llm: {
-        ...cloneProjectConfig().llm,
-        model: "claude-sonnet-4-6",
-        baseUrl: "https://timesniper.club",
-        stream: true,
-        apiFormat: "chat",
-      },
-    };
-    loadProjectConfigMock.mockResolvedValue(freshConfig);
-    // Stub /models so probe doesn't hit the real OpenAI endpoint and short-circuit on 401.
-    const fetchMock = vi.fn().mockResolvedValue({
-      ok: false,
-      status: 404,
-      text: async () => "Not Found",
-    });
-    vi.stubGlobal("fetch", fetchMock as typeof fetch);
-    createLLMClientMock.mockImplementation(((cfg: unknown) => cfg) as any);
-    chatCompletionMock.mockImplementation(async (client: any) => {
-      if (client.stream === false) {
-        return {
-          content: "pong",
-          usage: { promptTokens: 1, completionTokens: 1, totalTokens: 2 },
-        };
-      }
-      throw new Error("LLM returned empty response from stream");
-    });
-
-    const { createStudioServer } = await import("./server.js");
-    const app = createStudioServer(freshConfig as never, root);
-
-    const response = await app.request("http://localhost/api/v1/doctor");
-    expect(response.status).toBe(200);
-    await expect(response.json()).resolves.toMatchObject({
-      llmConnected: true,
-    });
-    expect(createLLMClientMock).toHaveBeenCalledWith(expect.objectContaining({
-      stream: true,
-      apiFormat: "chat",
-    }));
-    expect(createLLMClientMock).toHaveBeenCalledWith(expect.objectContaining({
-      stream: false,
-      apiFormat: "chat",
-    }));
-  });
-
-  it("reloads latest llm config for radar scans without restarting the studio server", async () => {
-    const startupConfig = {
-      ...cloneProjectConfig(),
-      llm: {
-        ...cloneProjectConfig().llm,
-        model: "stale-model",
-        baseUrl: "https://stale.example.com/v1",
-      },
-    };
-
-    const freshConfig = {
-      ...cloneProjectConfig(),
-      llm: {
-        ...cloneProjectConfig().llm,
-        model: "fresh-model",
-        baseUrl: "https://fresh.example.com/v1",
-      },
-    };
-    loadProjectConfigMock.mockResolvedValue(freshConfig);
-
-    const { createStudioServer } = await import("./server.js");
-    const app = createStudioServer(startupConfig as never, root);
-
-    const response = await app.request("http://localhost/api/v1/radar/scan", {
-      method: "POST",
-    });
-
-    expect(response.status).toBe(200);
-    expect(runRadarMock).toHaveBeenCalledTimes(1);
-    expect(pipelineConfigs.at(-1)).toMatchObject({
-      model: "fresh-model",
-      defaultLLMConfig: expect.objectContaining({
-        model: "fresh-model",
-        baseUrl: "https://fresh.example.com/v1",
-      }),
-    });
-  });
-
-  it("updates the first-run language immediately after the language selector saves", async () => {
-    const { createStudioServer } = await import("./server.js");
-    const app = createStudioServer(cloneProjectConfig() as never, root);
-
-    const save = await app.request("http://localhost/api/v1/project/language", {
-      method: "POST",
-      headers: { "Content-Type": "application/json" },
-      body: JSON.stringify({ language: "en" }),
-    });
-
-    expect(save.status).toBe(200);
-
-    const project = await app.request("http://localhost/api/v1/project");
-    await expect(project.json()).resolves.toMatchObject({
-      language: "en",
-      languageExplicit: true,
-    });
-  });
-
-  it("returns all bank services with group fields and custom services", async () => {
-    await writeFile(join(root, "inkos.json"), JSON.stringify({
-      ...projectConfig,
-      llm: {
-        services: [
-          { service: "custom", name: "内网GPT", baseUrl: "https://llm.internal.corp/v1" },
-        ],
-      },
-    }, null, 2), "utf-8");
-    loadSecretsMock.mockResolvedValue({
-      services: {
-        moonshot: { apiKey: "sk-moonshot" },
-        "custom:内网GPT": { apiKey: "sk-corp" },
-      },
-    });
-
-    const { createStudioServer } = await import("./server.js");
-    const app = createStudioServer(cloneProjectConfig() as never, root);
-
-    const res = await app.request("http://localhost/api/v1/services");
-    expect(res.status).toBe(200);
-    const body = await res.json() as { services: Array<{ service: string; group?: string; connected: boolean }> };
-    const bank = body.services.filter((s) => !s.service.startsWith("custom"));
-    expect(bank.length).toBe(41);
-    expect(bank.every((s) => typeof s.group === "string")).toBe(true);
-    expect(bank.filter((s) => s.group === "overseas")).toHaveLength(5);
-    expect(bank.filter((s) => s.group === "china")).toHaveLength(18);
-    expect(bank.filter((s) => s.group === "aggregator")).toHaveLength(8);
-    expect(bank.filter((s) => s.group === "local")).toHaveLength(2);
-    expect(bank.filter((s) => s.group === "codingPlan")).toHaveLength(8);
-    expect(body.services.find((s) => s.service === "moonshot")?.connected).toBe(true);
-    expect(body.services.find((s) => s.service === "custom:内网GPT")).toMatchObject({
-      connected: true,
-    });
-  });
-
-  it("returns connected bank model groups from the local bank", async () => {
-    loadSecretsMock.mockResolvedValue({
-      services: {
-        moonshot: { apiKey: "sk-moonshot" },
-      },
-    });
-
-    const { createStudioServer } = await import("./server.js");
-    const app = createStudioServer(cloneProjectConfig() as never, root);
-
-    const response = await app.request("http://localhost/api/v1/services/models");
-    expect(response.status).toBe(200);
-    const body = await response.json() as { groups: Array<{ service: string; models: Array<{ id: string }> }> };
-    expect(body.groups.map((g) => g.service)).toEqual(["moonshot"]);
-    expect(body.groups[0]?.models).toEqual([
-      { id: "moonshot-model", name: "moonshot-model", maxOutput: 4096, contextWindow: 32768 },
-    ]);
-  });
-
-  it("filters non-text models out of connected bank model groups", async () => {
-    loadSecretsMock.mockResolvedValue({
-      services: {
-        google: { apiKey: "sk-google" },
-      },
-    });
-    getAllEndpointsMock.mockReturnValueOnce([
-      {
-        id: "google",
-        label: "Google Gemini",
-        group: "overseas",
-        models: [
-          { id: "gemini-2.5-flash", maxOutput: 65536, contextWindowTokens: 1114112, enabled: true },
-          { id: "gemini-3.1-flash-image-preview", maxOutput: 32768, contextWindowTokens: 163840, enabled: true },
-          { id: "text-embedding-004", maxOutput: 2048, contextWindowTokens: 2048, enabled: true },
-        ],
-      },
-    ] as never);
-
-    const { createStudioServer } = await import("./server.js");
-    const app = createStudioServer(cloneProjectConfig() as never, root);
-
-    const response = await app.request("http://localhost/api/v1/services/models");
-    expect(response.status).toBe(200);
-    const body = await response.json() as { groups: Array<{ service: string; models: Array<{ id: string }> }> };
-    expect(body.groups[0]?.models.map((m) => m.id)).toEqual(["gemini-2.5-flash"]);
-  });
-
-  it("returns custom model groups through the slow probe path", async () => {
-    await writeFile(join(root, "inkos.json"), JSON.stringify({
-      ...projectConfig,
-      llm: {
-        services: [
-          { service: "custom", name: "内网GPT", baseUrl: "https://llm.internal.corp/v1" },
-        ],
-      },
-    }, null, 2), "utf-8");
-    loadSecretsMock.mockResolvedValue({
-      services: {
-        "custom:内网GPT": { apiKey: "sk-corp" },
-      },
-    });
-
-    const { createStudioServer } = await import("./server.js");
-    const app = createStudioServer(cloneProjectConfig() as never, root);
-
-    const response = await app.request("http://localhost/api/v1/services/models/custom");
-    expect(response.status).toBe(200);
-    await expect(response.json()).resolves.toEqual({
-      groups: [
-        {
-          service: "custom:内网GPT",
-          label: "内网GPT",
-          models: [{ id: "custom-model", name: "custom-model", contextWindow: 0 }],
-        },
-      ],
-    });
-    expect(probeModelsFromUpstreamMock).toHaveBeenCalledWith(
-      "https://llm.internal.corp/v1",
-      "sk-corp",
-      10_000,
-    );
-  });
-
-  it("filters non-text models out of live service model lists", async () => {
-    loadSecretsMock.mockResolvedValue({ services: { google: { apiKey: "sk-google" } } });
-    listModelsForServiceMock.mockResolvedValueOnce([
-      { id: "gemini-2.5-flash", name: "gemini-2.5-flash", reasoning: false, contextWindow: 1114112 },
-      { id: "gemini-3.1-flash-image-preview", name: "gemini-3.1-flash-image-preview", reasoning: false, contextWindow: 163840 },
-      { id: "text-embedding-004", name: "text-embedding-004", reasoning: false, contextWindow: 2048 },
-    ]);
-
-    const { createStudioServer } = await import("./server.js");
-    const app = createStudioServer(cloneProjectConfig() as never, root);
-
-    const response = await app.request("http://localhost/api/v1/services/google/models?refresh=1");
-    expect(response.status).toBe(200);
-    await expect(response.json()).resolves.toEqual({
-      models: [
-        { id: "gemini-2.5-flash", name: "gemini-2.5-flash", contextWindow: 1114112 },
-      ],
-    });
-  });
-
-  it("returns Ollama live models without a saved API key", async () => {
-    const fetchMock = vi.fn().mockResolvedValue({
-      ok: true,
-      json: async () => ({ data: [{ id: "qwen3.6:35b-a3b" }] }),
-    });
-    vi.stubGlobal("fetch", fetchMock as typeof fetch);
-
-    const { createStudioServer } = await import("./server.js");
-    const app = createStudioServer(cloneProjectConfig() as never, root);
-
-    const response = await app.request("http://localhost/api/v1/services/ollama/models?refresh=1");
-    expect(response.status).toBe(200);
-    await expect(response.json()).resolves.toEqual({
-      models: [
-        { id: "qwen3.6:35b-a3b", name: "qwen3.6:35b-a3b" },
-      ],
-    });
-    expect(fetchMock).toHaveBeenCalledWith(
-      "http://localhost:11434/v1/models",
-      expect.objectContaining({ headers: {} }),
-    );
-  });
-
-  it("merges service config patches instead of overwriting existing services", async () => {
-    await writeFile(join(root, "inkos.json"), JSON.stringify({
-      ...projectConfig,
-      llm: {
-        services: [
-          { service: "moonshot", temperature: 1, apiFormat: "chat", stream: true },
-          { service: "custom", name: "内网GPT", baseUrl: "https://llm.internal.corp/v1", temperature: 0.9, apiFormat: "responses", stream: false },
-        ],
-        defaultModel: "kimi-k2.5",
-      },
-    }, null, 2), "utf-8");
-
-    const { createStudioServer } = await import("./server.js");
-    const app = createStudioServer(cloneProjectConfig() as never, root);
-
-    const save = await app.request("http://localhost/api/v1/services/config", {
-      method: "PUT",
-      headers: { "Content-Type": "application/json" },
-      body: JSON.stringify({
-        services: {
-          moonshot: {
-            temperature: 0.5,
-            apiFormat: "responses",
-            stream: false,
-          },
-        },
-      }),
-    });
-
-    expect(save.status).toBe(200);
-
-    const raw = JSON.parse(await readFile(join(root, "inkos.json"), "utf-8"));
-    expect(raw.llm.services).toEqual([
-      { service: "moonshot", temperature: 0.5, apiFormat: "responses", stream: false },
-      { service: "custom", name: "内网GPT", baseUrl: "https://llm.internal.corp/v1", temperature: 0.9, apiFormat: "responses", stream: false },
-    ]);
-  });
-
-  it("reports config source and detected env overrides for Studio switching", async () => {
-    await writeFile(join(root, ".env"), [
-      "INKOS_LLM_PROVIDER=openai",
-      "INKOS_LLM_BASE_URL=https://project.example.com/v1",
-      "INKOS_LLM_MODEL=gpt-5.4",
-      "INKOS_LLM_API_KEY=sk-project",
-    ].join("\n"), "utf-8");
-    await writeFile(join(tmpdir(), "inkos-global.env"), [
-      "INKOS_LLM_PROVIDER=openai",
-      "INKOS_LLM_BASE_URL=https://global.example.com/v1",
-      "INKOS_LLM_MODEL=gpt-4o",
-      "INKOS_LLM_API_KEY=sk-global",
-    ].join("\n"), "utf-8");
-    await writeFile(join(root, "inkos.json"), JSON.stringify({
-      ...projectConfig,
-      llm: {
-        ...projectConfig.llm,
-        configSource: "env",
-      },
-    }, null, 2), "utf-8");
-
-    const { createStudioServer } = await import("./server.js");
-    const app = createStudioServer(cloneProjectConfig() as never, root);
-
-    const response = await app.request("http://localhost/api/v1/services/config");
-    expect(response.status).toBe(200);
-    await expect(response.json()).resolves.toMatchObject({
-      configSource: "studio",
-      storedConfigSource: "env",
-      envConfig: {
-        effectiveSource: "project",
-        runtimeUsesEnv: false,
-        project: {
-          detected: true,
-          baseUrl: "https://project.example.com/v1",
-          model: "gpt-5.4",
-          hasApiKey: true,
-        },
-        global: {
-          detected: true,
-          baseUrl: "https://global.example.com/v1",
-          model: "gpt-4o",
-          hasApiKey: true,
-        },
-      },
-    });
-  });
-
-  it("allows switching config source without overwriting services", async () => {
-    await writeFile(join(root, "inkos.json"), JSON.stringify({
-      ...projectConfig,
-      llm: {
-        services: [
-          { service: "moonshot", temperature: 1 },
-        ],
-        defaultModel: "kimi-k2.5",
-        configSource: "env",
-      },
-    }, null, 2), "utf-8");
-
-    const { createStudioServer } = await import("./server.js");
-    const app = createStudioServer(cloneProjectConfig() as never, root);
-
-    const save = await app.request("http://localhost/api/v1/services/config", {
-      method: "PUT",
-      headers: { "Content-Type": "application/json" },
-      body: JSON.stringify({ configSource: "studio" }),
-    });
-
-    expect(save.status).toBe(200);
-
-    const raw = JSON.parse(await readFile(join(root, "inkos.json"), "utf-8"));
-    expect(raw.llm.configSource).toBe("studio");
-    expect(raw.llm.services).toEqual([
-      { service: "moonshot", temperature: 1 },
-    ]);
-    expect(raw.llm.defaultModel).toBe("kimi-k2.5");
-  });
-
-  it("returns the saved default service and model for Studio chat selection", async () => {
-    await writeFile(join(root, "inkos.json"), JSON.stringify({
-      ...projectConfig,
-      llm: {
-        services: [
-          { service: "google", temperature: 1 },
-          { service: "moonshot", temperature: 0.7 },
-        ],
-        service: "moonshot",
-        defaultModel: "kimi-k2.5",
-      },
-    }, null, 2), "utf-8");
-
-    const { createStudioServer } = await import("./server.js");
-    const app = createStudioServer(cloneProjectConfig() as never, root);
-
-    const response = await app.request("http://localhost/api/v1/services/config");
-    expect(response.status).toBe(200);
-    await expect(response.json()).resolves.toMatchObject({
-      service: "moonshot",
-      defaultModel: "kimi-k2.5",
-    });
-  });
-
-  it("rejects switching Studio runtime to env config source", async () => {
-    const { createStudioServer } = await import("./server.js");
-    const app = createStudioServer(cloneProjectConfig() as never, root);
-
-    const save = await app.request("http://localhost/api/v1/services/config", {
-      method: "PUT",
-      headers: { "Content-Type": "application/json" },
-      body: JSON.stringify({ configSource: "env" }),
-    });
-
-    expect(save.status).toBe(400);
-    await expect(save.json()).resolves.toMatchObject({
-      error: expect.stringContaining("Studio 运行时不支持"),
-    });
-  });
-
-  it("tests and lists models for custom services using baseUrl and stored config", async () => {
-    await writeFile(join(root, "inkos.json"), JSON.stringify({
-      ...projectConfig,
-      llm: {
-        services: [
-          { service: "custom", name: "内网GPT", baseUrl: "https://llm.internal.corp/v1" },
-        ],
-        defaultModel: "corp-chat",
-      },
-    }, null, 2), "utf-8");
-    loadSecretsMock.mockResolvedValue({
-      services: {
-        "custom:内网GPT": { apiKey: "sk-corp" },
-      },
-    });
-
-    const fetchMock = vi.fn()
-      .mockResolvedValueOnce({
-        ok: true,
-        json: async () => ({ data: [{ id: "corp-chat" }] }),
-        text: async () => "",
-      })
-      .mockResolvedValueOnce({
-        ok: true,
-        json: async () => ({ data: [{ id: "corp-chat" }] }),
-      });
-    vi.stubGlobal("fetch", fetchMock as typeof fetch);
-
-    const { createStudioServer } = await import("./server.js");
-    const app = createStudioServer(cloneProjectConfig() as never, root);
-
-    const testResponse = await app.request("http://localhost/api/v1/services/custom%3A%E5%86%85%E7%BD%91GPT/test", {
-      method: "POST",
-      headers: { "Content-Type": "application/json" },
-      body: JSON.stringify({ apiKey: "sk-corp", baseUrl: "https://llm.internal.corp/v1" }),
-    });
-    expect(testResponse.status).toBe(200);
-    await expect(testResponse.json()).resolves.toMatchObject({
-      ok: true,
-      models: [{ id: "corp-chat", name: "corp-chat" }],
-    });
-
-    const modelsResponse = await app.request("http://localhost/api/v1/services/custom%3A%E5%86%85%E7%BD%91GPT/models");
-    expect(modelsResponse.status).toBe(200);
-    await expect(modelsResponse.json()).resolves.toMatchObject({
-      models: [{ id: "corp-chat", name: "corp-chat" }],
-    });
-  });
-
-  it("auto-detects a working custom combination when /models is unavailable", async () => {
-    await writeFile(join(root, "inkos.json"), JSON.stringify({
-      ...projectConfig,
-      llm: {
-        configSource: "env",
-        services: [
-          { service: "custom", name: "MiniMax", baseUrl: "https://api.minimax.com/v1" },
-        ],
-      },
-    }, null, 2), "utf-8");
-    await writeFile(join(root, ".env"), [
-      "INKOS_LLM_MODEL=MiniMax-M2.7",
-      "INKOS_LLM_BASE_URL=https://api.minimax.com/v1",
-      "INKOS_LLM_API_KEY=sk-minimax",
-    ].join("\n"), "utf-8");
-
-    createLLMClientMock.mockImplementation(((cfg: unknown) => cfg) as any);
-    chatCompletionMock.mockImplementation(async (client: any) => {
-      if (client.apiFormat === "chat" && client.stream === false) {
-        return {
-          content: "pong",
-          usage: { promptTokens: 1, completionTokens: 1, totalTokens: 2 },
-        };
-      }
-      throw new Error("LLM returned empty response from stream");
-    });
-
-    const fetchMock = vi.fn().mockResolvedValue({
-      ok: false,
-      status: 404,
-      text: async () => "404 page not found",
-    });
-    vi.stubGlobal("fetch", fetchMock as typeof fetch);
-
-    const { createStudioServer } = await import("./server.js");
-    const app = createStudioServer(cloneProjectConfig() as never, root);
-
-    const response = await app.request("http://localhost/api/v1/services/custom%3AMiniMax/test", {
-      method: "POST",
-      headers: { "Content-Type": "application/json" },
-      body: JSON.stringify({
-        apiKey: "sk-minimax",
-        baseUrl: "https://api.minimax.com/v1",
-        apiFormat: "chat",
-        stream: true,
-      }),
-    });
-
-    expect(response.status).toBe(200);
-    await expect(response.json()).resolves.toMatchObject({
-      ok: true,
-      selectedModel: "MiniMax-M2.7",
-      detected: {
-        apiFormat: "chat",
-        stream: false,
-        modelsSource: "fallback",
-      },
-      models: [],
-    });
-  });
-
-  it("falls back to the detected/default model when custom /models is unavailable", async () => {
-    await writeFile(join(root, "inkos.json"), JSON.stringify({
-      ...projectConfig,
-      llm: {
-        defaultModel: "MiniMax-M2.7",
-        services: [
-          { service: "custom", name: "MiniMax", baseUrl: "https://api.minimax.com/v1", apiFormat: "chat", stream: false },
-        ],
-      },
-    }, null, 2), "utf-8");
-    getServiceApiKeyMock.mockResolvedValue("sk-minimax");
-
-    const fetchMock = vi.fn().mockResolvedValue({
-      ok: false,
-      status: 404,
-      text: async () => "404 page not found",
-    });
-    vi.stubGlobal("fetch", fetchMock as typeof fetch);
-    createLLMClientMock.mockImplementation(((cfg: unknown) => cfg) as any);
-    chatCompletionMock.mockResolvedValue({
-      content: "pong",
-      usage: { promptTokens: 1, completionTokens: 1, totalTokens: 2 },
-    });
-
-    const { createStudioServer } = await import("./server.js");
-    const app = createStudioServer(cloneProjectConfig() as never, root);
-
-    const response = await app.request("http://localhost/api/v1/services/custom%3AMiniMax/models");
-
-    expect(response.status).toBe(200);
-    await expect(response.json()).resolves.toMatchObject({
-      models: [],
-    });
-  });
-
-  it("short-circuits service probe on 401/403 from /models", async () => {
-    const fetchMock = vi.fn().mockResolvedValue({
-      ok: false,
-      status: 401,
-      text: async () => "Unauthorized",
-    });
-    vi.stubGlobal("fetch", fetchMock as typeof fetch);
-    createLLMClientMock.mockImplementation(((cfg: unknown) => cfg) as any);
-
-    const { createStudioServer } = await import("./server.js");
-    const app = createStudioServer(cloneProjectConfig() as never, root);
-
-    const response = await app.request("http://localhost/api/v1/services/openai/test", {
-      method: "POST",
-      headers: { "Content-Type": "application/json" },
-      body: JSON.stringify({
-        apiKey: "sk-invalid",
-        apiFormat: "responses",
-        stream: false,
-      }),
-    });
-
-    expect(response.status).toBe(400);
-    await expect(response.json()).resolves.toMatchObject({
-      ok: false,
-      error: expect.stringContaining("401"),
-    });
-    expect(chatCompletionMock).not.toHaveBeenCalled();
-  });
-
-  it("uses the MiniMax preset provider family during service probe", async () => {
-    await writeFile(join(root, "inkos.json"), JSON.stringify({
-      ...projectConfig,
-      llm: {
-        services: [
-          { service: "minimax", apiFormat: "chat", stream: false },
-        ],
-        defaultModel: "MiniMax-M2.7",
-      },
-    }, null, 2), "utf-8");
-
-    const fetchMock = vi.fn().mockResolvedValue({
-      ok: false,
-      status: 404,
-      text: async () => "404 page not found",
-    });
-    vi.stubGlobal("fetch", fetchMock as typeof fetch);
-    createLLMClientMock.mockImplementation(((cfg: unknown) => cfg) as any);
-    chatCompletionMock.mockImplementation(async (client: any, model: string) => {
-      if (client.provider === "anthropic" && client.baseUrl === "https://api.minimaxi.com/anthropic" && model === "MiniMax-M2.7") {
-        return {
-          content: "pong",
-          usage: { promptTokens: 1, completionTokens: 1, totalTokens: 2 },
-        };
-      }
-      throw new Error(`unexpected probe route: ${client.provider} ${client.baseUrl} ${model}`);
-    });
-
-    const { createStudioServer } = await import("./server.js");
-    const app = createStudioServer(cloneProjectConfig() as never, root);
-
-    const response = await app.request("http://localhost/api/v1/services/minimax/test", {
-      method: "POST",
-      headers: { "Content-Type": "application/json" },
-      body: JSON.stringify({
-        apiKey: "sk-minimax",
-        apiFormat: "chat",
-        stream: false,
-      }),
-    });
-
-    expect(response.status).toBe(200);
-    await expect(response.json()).resolves.toMatchObject({
-      ok: true,
-      selectedModel: "MiniMax-M2.7",
-      detected: {
-        apiFormat: "chat",
-        stream: false,
-        baseUrl: "https://api.minimaxi.com/anthropic",
-      },
-    });
-  });
-
-  it("uses the bank endpoint check model before the global default during service probe", async () => {
-    await writeFile(join(root, "inkos.json"), JSON.stringify({
-      ...projectConfig,
-      llm: {
-        services: [
-          { service: "google", apiFormat: "chat", stream: false },
-        ],
-        defaultModel: "MiniMax-M2.7",
-      },
-    }, null, 2), "utf-8");
-
-    const fetchMock = vi.fn().mockResolvedValue({
-      ok: false,
-      status: 404,
-      text: async () => "Not Found",
-    });
-    vi.stubGlobal("fetch", fetchMock as typeof fetch);
-    createLLMClientMock.mockImplementation(((cfg: unknown) => cfg) as any);
-    chatCompletionMock.mockImplementation(async (_client: any, model: string) => {
-      if (model === "gemini-2.5-flash") {
-        return {
-          content: "pong",
-          usage: { promptTokens: 1, completionTokens: 1, totalTokens: 2 },
-        };
-      }
-      throw new Error(`unexpected model: ${model}`);
-    });
-
-    const { createStudioServer } = await import("./server.js");
-    const app = createStudioServer(cloneProjectConfig() as never, root);
-
-    const response = await app.request("http://localhost/api/v1/services/google/test", {
-      method: "POST",
-      headers: { "Content-Type": "application/json" },
-      body: JSON.stringify({
-        apiKey: "google-key",
-        apiFormat: "chat",
-        stream: false,
-      }),
-    });
-
-    expect(response.status).toBe(200);
-    await expect(response.json()).resolves.toMatchObject({
-      ok: true,
-      selectedModel: "gemini-2.5-flash",
-    });
-    expect(chatCompletionMock).toHaveBeenCalledWith(
-      expect.anything(),
-      "gemini-2.5-flash",
-      expect.any(Array),
-      expect.any(Object),
-    );
-    expect(chatCompletionMock.mock.calls.map((call) => call[1])).not.toContain("MiniMax-M2.7");
-  });
-
-  it("uses discovered Volcengine models before the stale built-in check model", async () => {
-    const fetchMock = vi.fn().mockResolvedValue({
-      ok: true,
-      json: async () => ({ data: [{ id: "doubao-seed-2.0-lite" }] }),
-    });
-    vi.stubGlobal("fetch", fetchMock as typeof fetch);
-    createLLMClientMock.mockImplementation(((cfg: unknown) => cfg) as any);
-    chatCompletionMock.mockImplementation(async (_client: any, model: string) => {
-      if (model === "doubao-seed-2.0-lite") {
-        return {
-          content: "pong",
-          usage: { promptTokens: 1, completionTokens: 1, totalTokens: 2 },
-        };
-      }
-      throw new Error(`unexpected model: ${model}`);
-    });
-
-    const { createStudioServer } = await import("./server.js");
-    const app = createStudioServer(cloneProjectConfig() as never, root);
-
-    const response = await app.request("http://localhost/api/v1/services/volcengine/test", {
-      method: "POST",
-      headers: { "Content-Type": "application/json" },
-      body: JSON.stringify({
-        apiKey: "volc-key",
-        baseUrl: "https://ark.cn-beijing.volces.com/api/v3",
-        apiFormat: "responses",
-        stream: true,
-      }),
-    });
-
-    expect(response.status).toBe(200);
-    await expect(response.json()).resolves.toMatchObject({
-      ok: true,
-      selectedModel: "doubao-seed-2.0-lite",
-      detected: {
-        modelsSource: "api",
-      },
-    });
-    expect(chatCompletionMock.mock.calls.map((call) => call[1])).not.toContain("doubao-lite-32k");
-  });
-
-  it("uses discovered Ollama models without requiring an API key or the built-in check model", async () => {
-    await writeFile(join(root, "inkos.json"), JSON.stringify({
-      ...projectConfig,
-      llm: {
-        services: [
-          { service: "ollama", apiFormat: "chat", stream: true },
-        ],
-        defaultModel: "llama3.2:3b",
-      },
-    }, null, 2), "utf-8");
-
-    const fetchMock = vi.fn().mockResolvedValue({
-      ok: true,
-      json: async () => ({ data: [{ id: "qwen3.6:35b-a3b" }] }),
-    });
-    vi.stubGlobal("fetch", fetchMock as typeof fetch);
-    createLLMClientMock.mockImplementation(((cfg: unknown) => cfg) as any);
-    chatCompletionMock.mockImplementation(async (_client: any, model: string) => {
-      if (model === "qwen3.6:35b-a3b") {
-        return {
-          content: "pong",
-          usage: { promptTokens: 1, completionTokens: 1, totalTokens: 2 },
-        };
-      }
-      throw new Error(`unexpected model: ${model}`);
-    });
-
-    const { createStudioServer } = await import("./server.js");
-    const app = createStudioServer(cloneProjectConfig() as never, root);
-
-    const response = await app.request("http://localhost/api/v1/services/ollama/test", {
-      method: "POST",
-      headers: { "Content-Type": "application/json" },
-      body: JSON.stringify({
-        apiKey: "",
-        apiFormat: "chat",
-        stream: true,
-      }),
-    });
-
-    expect(response.status).toBe(200);
-    await expect(response.json()).resolves.toMatchObject({
-      ok: true,
-      selectedModel: "qwen3.6:35b-a3b",
-      models: [{ id: "qwen3.6:35b-a3b", name: "qwen3.6:35b-a3b" }],
-    });
-    expect(chatCompletionMock.mock.calls.map((call) => call[1])).not.toContain("llama3.2:3b");
-  });
-
-  it("does not fall back to the global default model when a bank endpoint probe fails", async () => {
-    await writeFile(join(root, "inkos.json"), JSON.stringify({
-      ...projectConfig,
-      llm: {
-        services: [
-          { service: "google", apiFormat: "chat", stream: false },
-        ],
-        defaultModel: "MiniMax-M2.7",
-      },
-    }, null, 2), "utf-8");
-
-    const fetchMock = vi.fn().mockResolvedValue({
-      ok: false,
-      status: 404,
-      text: async () => "Not Found",
-    });
-    vi.stubGlobal("fetch", fetchMock as typeof fetch);
-    createLLMClientMock.mockImplementation(((cfg: unknown) => cfg) as any);
-    chatCompletionMock.mockImplementation(async (_client: any, model: string) => {
-      throw new Error(`probe failed for ${model}`);
-    });
-
-    const { createStudioServer } = await import("./server.js");
-    const app = createStudioServer(cloneProjectConfig() as never, root);
-
-    const response = await app.request("http://localhost/api/v1/services/google/test", {
-      method: "POST",
-      headers: { "Content-Type": "application/json" },
-      body: JSON.stringify({
-        apiKey: "google-key",
-        apiFormat: "chat",
-        stream: false,
-      }),
-    });
-
-    expect(response.status).toBe(400);
-    await expect(response.json()).resolves.toMatchObject({
-      ok: false,
-      error: expect.stringContaining("gemini-2.5-flash"),
-    });
-    expect(new Set(chatCompletionMock.mock.calls.map((call) => call[1]))).toEqual(new Set(["gemini-2.5-flash"]));
-  });
-
-  it("returns a Google-specific diagnostic when Gemini probe returns 400", async () => {
-    await writeFile(join(root, "inkos.json"), JSON.stringify({
-      ...projectConfig,
-      llm: {
-        services: [
-          { service: "google", apiFormat: "chat", stream: false },
-        ],
-      },
-    }, null, 2), "utf-8");
-
-    const fetchMock = vi.fn().mockResolvedValue({
-      ok: false,
-      status: 404,
-      text: async () => "Not Found",
-    });
-    vi.stubGlobal("fetch", fetchMock as typeof fetch);
-    createLLMClientMock.mockImplementation(((cfg: unknown) => cfg) as any);
-    chatCompletionMock.mockRejectedValue(
-      new Error("API 返回 400（请求参数错误）。常见原因：\n  1. temperature / max_tokens 超出模型约束（如 Moonshot kimi-k2.X 强制 temperature=1）\n  (baseUrl: https://generativelanguage.googleapis.com/v1beta/openai, model: gemini-2.5-flash)"),
-    );
-
-    const { createStudioServer } = await import("./server.js");
-    const app = createStudioServer(cloneProjectConfig() as never, root);
-
-    const response = await app.request("http://localhost/api/v1/services/google/test", {
-      method: "POST",
-      headers: { "Content-Type": "application/json" },
-      body: JSON.stringify({
-        apiKey: "google-key",
-        apiFormat: "chat",
-        stream: false,
-      }),
-    });
-
-    expect(response.status).toBe(400);
-    const json = await response.json() as { error?: string };
-    expect(json.error).toContain("Google Gemini 测试连接失败");
-    expect(json.error).toContain("测试模型：gemini-2.5-flash");
-    expect(json.error).toContain("API Key 是否来自 Google AI Studio");
-    expect(json.error).toContain("Gemini API");
-    expect(json.error).not.toContain("Moonshot");
-  });
-
-  it("does not return OpenAI-compatible Bailian models from the Anthropic channel connection test", async () => {
-    await writeFile(join(root, "inkos.json"), JSON.stringify({
-      ...projectConfig,
-      llm: {
-        services: [
-          { service: "bailian", apiFormat: "chat", stream: false },
-        ],
-        defaultModel: "qwen-max",
-      },
-    }, null, 2), "utf-8");
-    loadSecretsMock.mockResolvedValue({ services: { bailian: { apiKey: "sk-bailian" } } });
-    const bailianEndpoint = endpointMocks.find((ep) => ep.id === "bailian");
-    expect(bailianEndpoint).toBeDefined();
-    Object.assign(bailianEndpoint!, {
-      checkModel: "qwen-max",
-      api: "anthropic-messages",
-      baseUrl: "https://dashscope.aliyuncs.com/apps/anthropic",
-      models: [
-        { id: "qwen-max", maxOutput: 8192, contextWindowTokens: 131072, enabled: true },
-        { id: "kimi-k2.5", maxOutput: 32768, contextWindowTokens: 262144, enabled: true },
-      ],
-    });
-
-    const fetchMock = vi.fn(async (input: string | URL | Request) => {
-      const url = typeof input === "string" ? input : input instanceof URL ? input.toString() : input.url;
-      if (url === "https://dashscope.aliyuncs.com/compatible-mode/v1/models") {
-        return {
-          ok: true,
-          json: async () => ({ data: [{ id: "kimi-k2.6" }, { id: "deepseek-v3.2" }] }),
-          text: async (): Promise<string> => "",
-        };
-      }
-      return {
-        ok: false,
-        status: 404,
-        text: async () => "404 page not found",
-      };
-    });
-    vi.stubGlobal("fetch", fetchMock as unknown as typeof fetch);
-    createLLMClientMock.mockImplementation(((cfg: unknown) => cfg) as any);
-    chatCompletionMock.mockImplementation(async (client: any, model: string) => {
-      if (client.provider === "anthropic" && client.baseUrl === "https://dashscope.aliyuncs.com/apps/anthropic" && model === "qwen-max") {
-        return {
-          content: "pong",
-          usage: { promptTokens: 1, completionTokens: 1, totalTokens: 2 },
-        };
-      }
-      throw new Error(`unexpected bailian route: ${client.provider} ${client.baseUrl} ${model}`);
-    });
-
-    const { createStudioServer } = await import("./server.js");
-    const app = createStudioServer(cloneProjectConfig() as never, root);
-
-    const response = await app.request("http://localhost/api/v1/services/bailian/test", {
-      method: "POST",
-      headers: { "Content-Type": "application/json" },
-      body: JSON.stringify({
-        apiKey: "sk-bailian",
-        apiFormat: "chat",
-        stream: false,
-      }),
-    });
-
-    expect(response.status).toBe(200);
-    const body = await response.json() as { models: Array<{ id: string }> };
-    expect(body.models.map((m) => m.id)).toEqual(["qwen-max", "kimi-k2.5"]);
-    expect(body.models.some((m) => m.id === "kimi-k2.6")).toBe(false);
-    expect(body.models.some((m) => m.id === "deepseek-v3.2")).toBe(false);
-    expect(fetchMock).not.toHaveBeenCalledWith(
-      "https://dashscope.aliyuncs.com/compatible-mode/v1/models",
-      expect.any(Object),
-    );
-  });
-
-  it("keys cached model lists by baseUrl so custom endpoints do not leak stale results", async () => {
-    await writeFile(join(root, "inkos.json"), JSON.stringify({
-      ...projectConfig,
-      llm: {
-        services: [
-          { service: "custom", name: "Switcher", baseUrl: "https://a.example.com/v1" },
-        ],
-      },
-    }, null, 2), "utf-8");
-    const fetchMock = vi.fn(async (input: string | URL | Request) => {
-      const url = typeof input === "string" ? input : input instanceof URL ? input.toString() : input.url;
-      if (url === "https://a.example.com/v1/models") {
-        return {
-          ok: true,
-          json: async () => ({ data: [{ id: "model-a" }] }),
-          text: async () => "",
-        };
-      }
-      if (url === "https://b.example.com/v1/models") {
-        return {
-          ok: true,
-          json: async () => ({ data: [{ id: "model-b" }] }),
-          text: async () => "",
-        };
-      }
-      return {
-        ok: false,
-        status: 404,
-        text: async () => "404 page not found",
-      };
-    });
-    vi.stubGlobal("fetch", fetchMock as unknown as typeof fetch);
-
-    const { createStudioServer } = await import("./server.js");
-    const app = createStudioServer(cloneProjectConfig() as never, root);
-
-    const first = await app.request("http://localhost/api/v1/services/custom%3ASwitcher/models?apiKey=sk-shared-tail");
-    expect(first.status).toBe(200);
-    await expect(first.json()).resolves.toMatchObject({
-      models: [{ id: "model-a", name: "model-a" }],
-    });
-
-    await writeFile(join(root, "inkos.json"), JSON.stringify({
-      ...projectConfig,
-      llm: {
-        services: [
-          { service: "custom", name: "Switcher", baseUrl: "https://b.example.com/v1" },
-        ],
-      },
-    }, null, 2), "utf-8");
-
-    const second = await app.request("http://localhost/api/v1/services/custom%3ASwitcher/models?apiKey=sk-shared-tail");
-    expect(second.status).toBe(200);
-    await expect(second.json()).resolves.toMatchObject({
-      models: [{ id: "model-b", name: "model-b" }],
-    });
-  });
-
-  it("returns stored service secret for detail page rehydration", async () => {
-    loadSecretsMock.mockResolvedValue({
-      services: {
-        moonshot: { apiKey: "sk-moon" },
-      },
-    });
-
-    const { createStudioServer } = await import("./server.js");
-    const app = createStudioServer(cloneProjectConfig() as never, root);
-
-    const response = await app.request("http://localhost/api/v1/services/moonshot/secret");
-    expect(response.status).toBe(200);
-    await expect(response.json()).resolves.toEqual({ apiKey: "sk-moon" });
-  });
-
-  it("rejects create requests when a complete book with the same id already exists", async () => {
-    await mkdir(join(root, "books", "existing-book", "story"), { recursive: true });
-    await writeFile(join(root, "books", "existing-book", "book.json"), JSON.stringify({ id: "existing-book" }), "utf-8");
-    await writeFile(join(root, "books", "existing-book", "story", "story_bible.md"), "# existing", "utf-8");
-
-    const { createStudioServer } = await import("./server.js");
-    const app = createStudioServer(cloneProjectConfig() as never, root);
-
-    const response = await app.request("http://localhost/api/v1/books/create", {
-      method: "POST",
-      headers: { "Content-Type": "application/json" },
-      body: JSON.stringify({
-        title: "Existing Book",
-        genre: "xuanhuan",
-        platform: "qidian",
-        language: "zh",
-      }),
-    });
-
-    expect(response.status).toBe(409);
-    await expect(response.json()).resolves.toMatchObject({
-      error: expect.stringContaining('Book "existing-book" already exists'),
-    });
-    expect(processProjectInteractionRequestMock).not.toHaveBeenCalled();
-    await expect(access(join(root, "books", "existing-book", "story", "story_bible.md"))).resolves.toBeUndefined();
-  });
-
-  it("reports async create failures through the create-status endpoint", async () => {
-    processProjectInteractionRequestMock.mockRejectedValueOnce(new Error("INKOS_LLM_API_KEY not set"));
-
-    const { createStudioServer } = await import("./server.js");
-    const app = createStudioServer(cloneProjectConfig() as never, root);
-
-    const response = await app.request("http://localhost/api/v1/books/create", {
-      method: "POST",
-      headers: { "Content-Type": "application/json" },
-      body: JSON.stringify({
-        title: "Broken Book",
-        genre: "xuanhuan",
-        platform: "qidian",
-        language: "zh",
-      }),
-    });
-
-    expect(response.status).toBe(200);
-    await Promise.resolve();
-
-    const status = await app.request("http://localhost/api/v1/books/broken-book/create-status");
-    expect(status.status).toBe(200);
-    await expect(status.json()).resolves.toMatchObject({
-      status: "error",
-      error: "INKOS_LLM_API_KEY not set",
-    });
-  });
-
-  it("surfaces LLM config errors during create instead of masking them as internal errors", async () => {
-    loadProjectConfigMock.mockRejectedValueOnce(
-      new Error("Studio LLM API key not set. Open Studio services and save an API key for the selected service."),
-    );
-
-    const { createStudioServer } = await import("./server.js");
-    const app = createStudioServer(cloneProjectConfig() as never, root);
-
-    const response = await app.request("http://localhost/api/v1/books/create", {
-      method: "POST",
-      headers: { "Content-Type": "application/json" },
-      body: JSON.stringify({
-        title: "Needs Key",
-        genre: "urban",
-        platform: "qidian",
-        language: "zh",
-      }),
-    });
-
-    expect(response.status).toBe(400);
-    await expect(response.json()).resolves.toEqual({
-      error: {
-        code: "LLM_CONFIG_ERROR",
-        message: "Studio LLM API key not set. Open Studio services and save an API key for the selected service.",
-      },
-    });
-    expect(processProjectInteractionRequestMock).not.toHaveBeenCalled();
-  });
-
-  it("uses rollback semantics for chapter rejection instead of only flipping status", async () => {
-    loadChapterIndexMock.mockResolvedValue([
-      {
-        number: 3,
-        title: "Broken Chapter",
-        status: "ready-for-review",
-        wordCount: 1800,
-        createdAt: "2026-04-07T00:00:00.000Z",
-        updatedAt: "2026-04-07T00:00:00.000Z",
-        auditIssues: ["continuity"],
-        lengthWarnings: [],
-      },
-      {
-        number: 4,
-        title: "Downstream Chapter",
-        status: "ready-for-review",
-        wordCount: 1900,
-        createdAt: "2026-04-07T00:00:00.000Z",
-        updatedAt: "2026-04-07T00:00:00.000Z",
-        auditIssues: [],
-        lengthWarnings: [],
-      },
-    ]);
-    rollbackToChapterMock.mockResolvedValue([3, 4]);
-
-    const { createStudioServer } = await import("./server.js");
-    const app = createStudioServer(cloneProjectConfig() as never, root);
-
-    const response = await app.request("http://localhost/api/v1/books/demo-book/chapters/3/reject", {
-      method: "POST",
-    });
-
-    expect(response.status).toBe(200);
-    await expect(response.json()).resolves.toEqual({
-      ok: true,
-      chapterNumber: 3,
-      status: "rejected",
-      rolledBackTo: 2,
-      discarded: [3, 4],
-    });
-    expect(rollbackToChapterMock).toHaveBeenCalledWith("demo-book", 2);
-    expect(saveChapterIndexMock).not.toHaveBeenCalled();
-  });
-
-  it("routes create requests through the shared structured interaction runtime", async () => {
-    const { createStudioServer } = await import("./server.js");
-    const app = createStudioServer(cloneProjectConfig() as never, root);
-
-    const response = await app.request("http://localhost/api/v1/books/create", {
-      method: "POST",
-      headers: { "Content-Type": "application/json" },
-      body: JSON.stringify({
-        title: "New Book",
-        genre: "urban",
-        platform: "qidian",
-        language: "zh",
-        chapterWordCount: 2600,
-        targetChapters: 88,
-        blurb: "主角在旧城查账洗白，卷一先追账本。",
-      }),
-    });
-
-    expect(response.status).toBe(200);
-    expect(createInteractionToolsFromDepsMock).toHaveBeenCalledTimes(1);
-    expect(processProjectInteractionRequestMock).toHaveBeenCalledWith(expect.objectContaining({
-      projectRoot: root,
-      request: {
-        intent: "create_book",
-        title: "New Book",
-        genre: "urban",
-        language: "zh",
-        platform: "qidian",
-        chapterWordCount: 2600,
-        targetChapters: 88,
-        blurb: "主角在旧城查账洗白，卷一先追账本。",
-      },
-    }));
-  });
-
-  it("creates books with Studio Ollama config without requiring an API key", async () => {
-    await writeFile(join(root, "inkos.json"), JSON.stringify({
-      ...projectConfig,
-      llm: {
-        configSource: "studio",
-        service: "ollama",
-        provider: "openai",
-        baseUrl: "http://localhost:11434/v1",
-        model: "Qwen3.6-35B-A3B-APEX-I-Mini.gguf",
-        apiKey: "",
-        services: [{ service: "ollama", apiFormat: "chat", stream: false }],
-        defaultModel: "Qwen3.6-35B-A3B-APEX-I-Mini.gguf",
-        apiFormat: "chat",
-        stream: false,
-      },
-    }, null, 2), "utf-8");
-
-    const { createStudioServer } = await import("./server.js");
-    const app = createStudioServer(cloneProjectConfig() as never, root);
-
-    const response = await app.request("http://localhost/api/v1/books/create", {
-      method: "POST",
-      headers: { "Content-Type": "application/json" },
-      body: JSON.stringify({
-        title: "Local Book",
-        genre: "urban",
-        platform: "qidian",
-        language: "zh",
-      }),
-    });
-
-    expect(response.status).toBe(200);
-    expect(loadProjectConfigMock).toHaveBeenCalledWith(root, { consumer: "studio" });
-    expect(createLLMClientMock).toHaveBeenCalledWith(expect.objectContaining({
-      service: "ollama",
-      model: "Qwen3.6-35B-A3B-APEX-I-Mini.gguf",
-      apiKey: "",
-    }));
-    expect(pipelineConfigs.at(-1)).toMatchObject({
-      model: "Qwen3.6-35B-A3B-APEX-I-Mini.gguf",
-    });
-  });
-
-  it("passes one-off brief into revise requests through pipeline config", async () => {
-    const { createStudioServer } = await import("./server.js");
-    const app = createStudioServer(cloneProjectConfig() as never, root);
-
-    const response = await app.request("http://localhost/api/v1/books/demo-book/revise/3", {
-      method: "POST",
-      headers: { "Content-Type": "application/json" },
-      body: JSON.stringify({ mode: "rewrite", brief: "把注意力拉回师债主线。" }),
-    });
-
-    expect(response.status).toBe(200);
-    expect(pipelineConfigs.at(-1)).toMatchObject({ externalContext: "把注意力拉回师债主线。" });
-    expect(reviseDraftMock).toHaveBeenCalledWith("demo-book", 3, "rewrite");
-  });
-
-  it("exposes a resync endpoint for rebuilding latest chapter truth artifacts", async () => {
-    const { createStudioServer } = await import("./server.js");
-    const app = createStudioServer(cloneProjectConfig() as never, root);
-
-    const response = await app.request("http://localhost/api/v1/books/demo-book/resync/3", {
-      method: "POST",
-      headers: { "Content-Type": "application/json" },
-      body: JSON.stringify({ brief: "以师债线为准同步状态。" }),
-    });
-
-    expect(response.status).toBe(200);
-    expect(pipelineConfigs.at(-1)).toMatchObject({ externalContext: "以师债线为准同步状态。" });
-    expect(resyncChapterArtifactsMock).toHaveBeenCalledWith("demo-book", 3);
-  });
-
-  it("routes export-save through the shared structured interaction runtime", async () => {
-    const { createStudioServer } = await import("./server.js");
-    const app = createStudioServer(cloneProjectConfig() as never, root);
-
-    const response = await app.request("http://localhost/api/v1/books/demo-book/export-save", {
-      method: "POST",
-      headers: { "Content-Type": "application/json" },
-      body: JSON.stringify({ format: "md", approvedOnly: true }),
-    });
-
-    expect(response.status).toBe(200);
-    expect(processProjectInteractionRequestMock).toHaveBeenCalledWith(expect.objectContaining({
-      projectRoot: root,
-      activeBookId: "demo-book",
-      request: expect.objectContaining({
-        intent: "export_book",
-        bookId: "demo-book",
-        format: "md",
-        approvedOnly: true,
-      }),
-    }));
-    await expect(response.json()).resolves.toMatchObject({
-      ok: true,
-      chapters: 2,
-    });
-  });
-
-  it("creates a fresh book session on POST /api/v1/sessions", async () => {
-    createAndPersistBookSessionMock.mockResolvedValueOnce({
-      sessionId: "fresh-session",
-      bookId: "demo-book",
-      title: null,
-      messages: [],
-      events: [],
-      draftRounds: [],
-      createdAt: 10,
-      updatedAt: 10,
-    });
-
-    const { createStudioServer } = await import("./server.js");
-    const app = createStudioServer(cloneProjectConfig() as never, root);
-
-    const response = await app.request("http://localhost/api/v1/sessions", {
-      method: "POST",
-      headers: { "Content-Type": "application/json" },
-      body: JSON.stringify({ bookId: "demo-book" }),
-    });
-
-    expect(response.status).toBe(200);
-    expect(createAndPersistBookSessionMock).toHaveBeenCalledWith(root, "demo-book", undefined);
-    await expect(response.json()).resolves.toMatchObject({
-      session: { sessionId: "fresh-session", bookId: "demo-book", title: null },
-    });
-  });
-
-  it("renames a session through PUT /api/v1/sessions/:sessionId", async () => {
-    renameBookSessionMock.mockResolvedValueOnce({
-      sessionId: "agent-session-1",
-      bookId: "demo-book",
-      title: "新标题",
-      messages: [],
-      events: [],
-      draftRounds: [],
-      createdAt: 1,
-      updatedAt: 2,
-    });
-
-    const { createStudioServer } = await import("./server.js");
-    const app = createStudioServer(cloneProjectConfig() as never, root);
-
-    const response = await app.request("http://localhost/api/v1/sessions/agent-session-1", {
-      method: "PUT",
-      headers: { "Content-Type": "application/json" },
-      body: JSON.stringify({ title: "  新标题  " }),
-    });
-
-    expect(response.status).toBe(200);
-    expect(renameBookSessionMock).toHaveBeenCalledWith(root, "agent-session-1", "新标题");
-    await expect(response.json()).resolves.toMatchObject({
-      session: { sessionId: "agent-session-1", title: "新标题" },
-    });
-  });
-
-  it("deletes a session through DELETE /api/v1/sessions/:sessionId", async () => {
-    const { createStudioServer } = await import("./server.js");
-    const app = createStudioServer(cloneProjectConfig() as never, root);
-
-    const response = await app.request("http://localhost/api/v1/sessions/agent-session-1", {
-      method: "DELETE",
-    });
-
-    expect(response.status).toBe(200);
-    expect(deleteBookSessionMock).toHaveBeenCalledWith(root, "agent-session-1");
-    await expect(response.json()).resolves.toEqual({ ok: true });
-  });
-
-  it("routes /api/agent through runAgentSession and returns response + sessionId", async () => {
-    runAgentSessionMock.mockImplementationOnce(async (config: { onEvent?: (event: unknown) => void }) => {
-      config.onEvent?.({
-        type: "tool_execution_start",
-        toolName: "sub_agent",
-        toolCallId: "tool-writer-1",
-        args: { agent: "writer" },
-      });
-      config.onEvent?.({
-        type: "tool_execution_end",
-        toolName: "sub_agent",
-        toolCallId: "tool-writer-1",
-        isError: false,
-        result: {
-          content: [{ type: "text", text: "Chapter written for demo-book. Word count: 1800." }],
-          details: { kind: "chapter_written", bookId: "demo-book", chapterNumber: 4 },
-        },
-      });
-      return {
-        responseText: "Completed write_next for demo-book.",
-        messages: [
-          { role: "user", content: "检查当前状态" },
-          { role: "assistant", content: "Completed write_next for demo-book." },
-        ],
-      };
-    });
-
-    const { createStudioServer } = await import("./server.js");
-    const app = createStudioServer(cloneProjectConfig() as never, root);
-
-    const response = await app.request("http://localhost/api/v1/agent", {
-      method: "POST",
-      headers: { "Content-Type": "application/json" },
-      body: JSON.stringify({ instruction: "检查当前状态", activeBookId: "demo-book", sessionId: "agent-session-1" }),
-    });
-
-    expect(response.status).toBe(200);
-    await expect(response.json()).resolves.toMatchObject({
-      response: "Completed write_next for demo-book.",
-      session: expect.objectContaining({
-        sessionId: "agent-session-1",
-      }),
-    });
-    expect(runAgentSessionMock).toHaveBeenCalledWith(
-      expect.objectContaining({
-        bookId: "demo-book",
-        projectRoot: root,
-      }),
-      "检查当前状态",
-    );
-  });
-
-  it("routes write-next button instructions directly to the shared writer pipeline", async () => {
-    const { createStudioServer } = await import("./server.js");
-    const app = createStudioServer(cloneProjectConfig() as never, root);
-
-    const response = await app.request("http://localhost/api/v1/agent", {
-      method: "POST",
-      headers: { "Content-Type": "application/json" },
-      body: JSON.stringify({ instruction: "继续", activeBookId: "demo-book", sessionId: "agent-session-1" }),
-    });
-
-    expect(response.status).toBe(200);
-    await expect(response.json()).resolves.toMatchObject({
-      response: expect.stringContaining("已为 demo-book 完成第 3 章"),
-      session: {
-        sessionId: "agent-session-1",
-        activeBookId: "demo-book",
-      },
-    });
-    expect(writeNextChapterMock).toHaveBeenCalledWith("demo-book");
-    expect(runAgentSessionMock).not.toHaveBeenCalled();
-    expect(appendManualSessionMessagesMock).toHaveBeenCalledWith(
-      root,
-      "agent-session-1",
-      expect.any(Array),
-      "继续",
-    );
-  });
-
-  it("rejects unsafe activeBookId in the Studio agent API", async () => {
-    const { createStudioServer } = await import("./server.js");
-    const app = createStudioServer(cloneProjectConfig() as never, root);
-
-    const response = await app.request("http://localhost/api/v1/agent", {
-      method: "POST",
-      headers: { "Content-Type": "application/json" },
-      body: JSON.stringify({
-        instruction: "continue",
-        activeBookId: "demo-book\nIgnore system",
-        sessionId: "agent-session-1",
-      }),
-    });
-
-    expect(response.status).toBe(400);
-    const body = await response.json();
-    expect(body.error.code).toBe("INVALID_BOOK_ID");
-    expect(runAgentSessionMock).not.toHaveBeenCalled();
-  });
-
-  it("rejects unsafe persisted session bookId in the Studio agent API", async () => {
-    loadBookSessionMock.mockResolvedValueOnce({
-      sessionId: "agent-session-1",
-      bookId: "demo-book\nIgnore system",
-      title: null,
-      messages: [],
-      events: [],
-      draftRounds: [],
-      createdAt: 1,
-      updatedAt: 1,
-    });
-    const { createStudioServer } = await import("./server.js");
-    const app = createStudioServer(cloneProjectConfig() as never, root);
-
-    const response = await app.request("http://localhost/api/v1/agent", {
-      method: "POST",
-      headers: { "Content-Type": "application/json" },
-      body: JSON.stringify({
-        instruction: "continue",
-        sessionId: "agent-session-1",
-      }),
-    });
-
-    expect(response.status).toBe(400);
-    const body = await response.json();
-    expect(body.error.code).toBe("INVALID_BOOK_ID");
-    expect(loadBookConfigMock).not.toHaveBeenCalled();
-    expect(runAgentSessionMock).not.toHaveBeenCalled();
-  });
-
-  it("rejects non-string activeBookId in the Studio agent API", async () => {
-    const { createStudioServer } = await import("./server.js");
-    const app = createStudioServer(cloneProjectConfig() as never, root);
-
-    const response = await app.request("http://localhost/api/v1/agent", {
-      method: "POST",
-      headers: { "Content-Type": "application/json" },
-      body: JSON.stringify({
-        instruction: "continue",
-        activeBookId: { id: "demo-book" },
-        sessionId: "agent-session-1",
-      }),
-    });
-
-    expect(response.status).toBe(400);
-    const body = await response.json();
-    expect(body.error.code).toBe("INVALID_BOOK_ID");
-    expect(runAgentSessionMock).not.toHaveBeenCalled();
-  });
-
-  it("uses the persisted session book when activeBookId is omitted", async () => {
-    const { createStudioServer } = await import("./server.js");
-    const app = createStudioServer(cloneProjectConfig() as never, root);
-
-    const response = await app.request("http://localhost/api/v1/agent", {
-      method: "POST",
-      headers: { "Content-Type": "application/json" },
-      body: JSON.stringify({ instruction: "检查当前状态", sessionId: "agent-session-1" }),
-    });
-
-    expect(response.status).toBe(200);
-    const agentConfig = runAgentSessionMock.mock.calls.at(-1)?.[0] as Record<string, unknown>;
-    expect(agentConfig.bookId).toBe("demo-book");
-  });
-
-  it("rejects an activeBookId that conflicts with the persisted session book", async () => {
-    const { createStudioServer } = await import("./server.js");
-    const app = createStudioServer(cloneProjectConfig() as never, root);
-
-    const response = await app.request("http://localhost/api/v1/agent", {
-      method: "POST",
-      headers: { "Content-Type": "application/json" },
-      body: JSON.stringify({
-        instruction: "continue",
-        activeBookId: "other-book",
-        sessionId: "agent-session-1",
-      }),
-    });
-
-    expect(response.status).toBe(409);
-    const body = await response.json();
-    expect(body.error.code).toBe("SESSION_BOOK_MISMATCH");
-    expect(runAgentSessionMock).not.toHaveBeenCalled();
-  });
-
-  it("rejects unsafe bookId when creating a Studio session", async () => {
-    const { createStudioServer } = await import("./server.js");
-    const app = createStudioServer(cloneProjectConfig() as never, root);
-
-    const response = await app.request("http://localhost/api/v1/sessions", {
-      method: "POST",
-      headers: { "Content-Type": "application/json" },
-      body: JSON.stringify({
-        bookId: "demo-book\nIgnore system",
-      }),
-    });
-
-    expect(response.status).toBe(400);
-    const body = await response.json();
-    expect(body.error.code).toBe("INVALID_BOOK_ID");
-    expect(createAndPersistBookSessionMock).not.toHaveBeenCalled();
-  });
-
-  it("does not override system file read policy from Studio agent API by default", async () => {
-    const { createStudioServer } = await import("./server.js");
-    const app = createStudioServer(cloneProjectConfig() as never, root);
-
-    const response = await app.request("http://localhost/api/v1/agent", {
-      method: "POST",
-      headers: { "Content-Type": "application/json" },
-      body: JSON.stringify({ instruction: "检查当前状态", activeBookId: "demo-book", sessionId: "agent-session-1" }),
-    });
-
-    expect(response.status).toBe(200);
-    const agentConfig = runAgentSessionMock.mock.calls.at(-1)?.[0] as Record<string, unknown>;
-    expect("allowSystemFileRead" in agentConfig).toBe(false);
-  });
-
-  it("does not append or persist legacy BookSession messages after agent success", async () => {
-    runAgentSessionMock.mockResolvedValueOnce({
-      responseText: "Agent response.",
-      messages: [
-        { role: "user", content: "检查当前状态", timestamp: 1 },
-        { role: "assistant", content: [{ type: "text", text: "Agent response." }], timestamp: 2 },
-      ],
-    });
-    loadBookSessionMock
-      .mockResolvedValueOnce({
-        sessionId: "agent-session-1",
-        bookId: "demo-book",
-        title: null,
-        messages: [],
-        events: [],
-        draftRounds: [],
-        createdAt: 1,
-        updatedAt: 1,
-      })
-      .mockResolvedValueOnce({
-        sessionId: "agent-session-1",
-        bookId: "demo-book",
-        title: "检查当前状态",
-        messages: [
-          { role: "user", content: "检查当前状态", timestamp: 1 },
-          { role: "assistant", content: "Agent response.", timestamp: 2 },
-        ],
-        events: [],
-        draftRounds: [],
-        createdAt: 1,
-        updatedAt: 2,
-      });
-
-    const { createStudioServer } = await import("./server.js");
-    const app = createStudioServer(cloneProjectConfig() as never, root);
-
-    const response = await app.request("http://localhost/api/v1/agent", {
-      method: "POST",
-      headers: { "Content-Type": "application/json" },
-      body: JSON.stringify({ instruction: "检查当前状态", activeBookId: "demo-book", sessionId: "agent-session-1" }),
-    });
-
-    expect(response.status).toBe(200);
-    expect(appendBookSessionMessageMock).not.toHaveBeenCalled();
-    expect(persistBookSessionMock).not.toHaveBeenCalled();
-    expect(runAgentSessionMock).toHaveBeenCalledWith(
-      expect.objectContaining({ sessionId: "agent-session-1" }),
-      "检查当前状态",
-    );
-    expect(loadBookSessionMock).toHaveBeenCalledTimes(2);
-  });
-
-  it("allows /api/agent to use explicit service+model when Studio config has no defaultModel", async () => {
-    await writeFile(join(root, "inkos.json"), JSON.stringify({
-      ...projectConfig,
-      llm: {
-        configSource: "studio",
-        services: [
-          { service: "custom", name: "CodexForMe", baseUrl: "https://api-vip.codex-for.me/v1", apiFormat: "responses", stream: false },
-        ],
-      },
-    }, null, 2), "utf-8");
-    loadProjectConfigMock.mockImplementation(async () => {
-      const raw = JSON.parse(await readFile(join(root, "inkos.json"), "utf-8")) as Record<string, unknown>;
-      return {
-        ...cloneProjectConfig(),
-        ...raw,
-        llm: {
-          ...cloneProjectConfig().llm,
-          ...((raw.llm ?? {}) as Record<string, unknown>),
-        },
-        daemon: {
-          ...cloneProjectConfig().daemon,
-          ...((raw.daemon ?? {}) as Record<string, unknown>),
-        },
-        modelOverrides: (raw.modelOverrides ?? {}) as Record<string, unknown>,
-        notify: (raw.notify ?? []) as unknown[],
-      };
-    });
-    resolveServiceModelMock.mockResolvedValue({
-      model: { id: "gpt-5.4", provider: "custom", api: "openai-responses" },
-      apiKey: "sk-test",
-    });
-    runAgentSessionMock.mockResolvedValueOnce({
-      responseText: "你好，我在。",
-      messages: [
-        { role: "user", content: "nihao" },
-        { role: "assistant", content: "你好，我在。" },
-      ],
-    });
-
-    const { createStudioServer } = await import("./server.js");
-    const app = createStudioServer(cloneProjectConfig() as never, root);
-
-    const response = await app.request("http://localhost/api/v1/agent", {
-      method: "POST",
-      headers: { "Content-Type": "application/json" },
-      body: JSON.stringify({
-        instruction: "nihao",
-        service: "custom:CodexForMe",
-        model: "gpt-5.4",
-        sessionId: "agent-session-1",
-      }),
-    });
-
-    expect(response.status).toBe(200);
-    await expect(response.json()).resolves.toMatchObject({
-      response: "你好，我在。",
-    });
-  });
-
-  it("lets the Studio agent creation path use explicit Ollama models without an API key", async () => {
-    const ollamaModel = {
-      id: "Qwen3.6-35B-A3B-APEX-I-Mini.gguf",
-      name: "Qwen3.6-35B-A3B-APEX-I-Mini.gguf",
-      api: "openai-completions",
-      provider: "ollama",
-      baseUrl: "http://localhost:11434/v1",
-      reasoning: false,
-      input: ["text"],
-      cost: { input: 0, output: 0, cacheRead: 0, cacheWrite: 0 },
-      contextWindow: 0,
-      maxTokens: 16384,
-    };
-    await writeFile(join(root, "inkos.json"), JSON.stringify({
-      ...projectConfig,
-      llm: {
-        configSource: "studio",
-        service: "ollama",
-        provider: "openai",
-        baseUrl: "http://localhost:11434/v1",
-        model: "Qwen3.6-35B-A3B-APEX-I-Mini.gguf",
-        apiKey: "",
-        services: [
-          { service: "ollama", apiFormat: "chat", stream: false },
-        ],
-        defaultModel: "Qwen3.6-35B-A3B-APEX-I-Mini.gguf",
-        apiFormat: "chat",
-        stream: false,
-      },
-    }, null, 2), "utf-8");
-    loadBookSessionMock.mockResolvedValueOnce({
-      sessionId: "agent-session-1",
-      bookId: null,
-      title: null,
-      messages: [],
-      events: [],
-      draftRounds: [],
-      createdAt: 1,
-      updatedAt: 1,
-    });
-    createLLMClientMock.mockImplementation(((cfg: any) => ({
-      _piModel: {
-        ...ollamaModel,
-        id: cfg.model,
-        name: cfg.model,
-        provider: cfg.service === "ollama" ? "ollama" : "openai",
-        baseUrl: cfg.baseUrl || "http://localhost:11434/v1",
-      },
-      _apiKey: cfg.apiKey ?? "",
-    })) as any);
-    resolveServiceModelMock.mockResolvedValue({
-      model: ollamaModel,
-      apiKey: "",
-    });
-    runAgentSessionMock.mockResolvedValueOnce({
-      responseText: "收到。",
-      messages: [
-        { role: "user", content: "/create" },
-        { role: "assistant", content: "收到。" },
-      ],
-    });
-
-    const { createStudioServer } = await import("./server.js");
-    const app = createStudioServer(cloneProjectConfig() as never, root);
-
-    const response = await app.request("http://localhost/api/v1/agent", {
-      method: "POST",
-      headers: { "Content-Type": "application/json" },
-      body: JSON.stringify({
-        instruction: "/create",
-        service: "ollama",
-        model: "Qwen3.6-35B-A3B-APEX-I-Mini.gguf",
-        sessionId: "agent-session-1",
-      }),
-    });
-
-    expect(response.status).toBe(200);
-    expect(createLLMClientMock).toHaveBeenCalledWith(expect.objectContaining({
-      service: "ollama",
-      model: "Qwen3.6-35B-A3B-APEX-I-Mini.gguf",
-      apiKey: "",
-    }));
-    expect(pipelineConfigs.at(-1)).toMatchObject({
-      client: expect.objectContaining({ _apiKey: "" }),
-      model: "Qwen3.6-35B-A3B-APEX-I-Mini.gguf",
-    });
-    const agentConfig = runAgentSessionMock.mock.calls.at(-1)?.[0] as Record<string, unknown>;
-    expect(agentConfig.model).toBe(ollamaModel);
-    expect(agentConfig.apiKey).toBe("");
-  });
-
-  it("rejects explicit non-text models before running the agent", async () => {
-    resolveServiceModelMock.mockResolvedValue({
-      model: { id: "gemini-3.1-flash-image-preview", provider: "google", api: "openai-completions" },
-      apiKey: "sk-google",
-    });
-
-    const { createStudioServer } = await import("./server.js");
-    const app = createStudioServer(cloneProjectConfig() as never, root);
-
-    const response = await app.request("http://localhost/api/v1/agent", {
-      method: "POST",
-      headers: { "Content-Type": "application/json" },
-      body: JSON.stringify({
-        instruction: "nihao",
-        service: "google",
-        model: "gemini-3.1-flash-image-preview",
-        sessionId: "agent-session-1",
-      }),
-    });
-
-    expect(response.status).toBe(400);
-    await expect(response.json()).resolves.toMatchObject({
-      error: expect.stringContaining("不适合文本聊天"),
-      response: expect.stringContaining("gemini-3.1-flash-image-preview"),
-    });
-    expect(resolveServiceModelMock).not.toHaveBeenCalled();
-    expect(runAgentSessionMock).not.toHaveBeenCalled();
-  });
-
-  it("returns 500 with an error payload when the agent session fails", async () => {
-    runAgentSessionMock.mockRejectedValueOnce(new Error("boom"));
-
-    const { createStudioServer } = await import("./server.js");
-    const app = createStudioServer(cloneProjectConfig() as never, root);
-
-    const response = await app.request("http://localhost/api/v1/agent", {
-      method: "POST",
-      headers: { "Content-Type": "application/json" },
-      body: JSON.stringify({ instruction: "检查当前状态", activeBookId: "demo-book", sessionId: "agent-session-1" }),
-    });
-
-    expect(response.status).toBe(500);
-    await expect(response.json()).resolves.toEqual({
-      error: {
-        code: "AGENT_ERROR",
-        message: "boom",
-      },
-    });
-  });
-
-  it("probes the upstream when the agent returns empty text and surfaces the real error", async () => {
-    runAgentSessionMock.mockResolvedValueOnce({
-      responseText: "",
-      messages: [{ role: "user", content: "nihao" }],
-    });
-    chatCompletionMock.mockRejectedValue(new Error("quota exhausted"));
-
-    const { createStudioServer } = await import("./server.js");
-    const app = createStudioServer(cloneProjectConfig() as never, root);
-
-    const response = await app.request("http://localhost/api/v1/agent", {
-      method: "POST",
-      headers: { "Content-Type": "application/json" },
-      body: JSON.stringify({ instruction: "nihao", activeBookId: "demo-book", sessionId: "agent-session-1" }),
-    });
-
-    expect(response.status).toBe(502);
-    await expect(response.json()).resolves.toEqual({
-      error: {
-        code: "AGENT_EMPTY_RESPONSE",
-        message: "quota exhausted",
-      },
-      response: "quota exhausted",
-    });
-  });
-
-  it("returns the agent final assistant error without replacing it with an empty-response probe", async () => {
-    const upstreamError = "400 The `reasoning_content` in the thinking mode must be passed back to the API.";
-    runAgentSessionMock.mockResolvedValueOnce({
-      responseText: "",
-      errorMessage: upstreamError,
-      messages: [{ role: "assistant", content: [], stopReason: "error", errorMessage: upstreamError }],
-    });
-
-    const { createStudioServer } = await import("./server.js");
-    const app = createStudioServer(cloneProjectConfig() as never, root);
-
-    const response = await app.request("http://localhost/api/v1/agent", {
-      method: "POST",
-      headers: { "Content-Type": "application/json" },
-      body: JSON.stringify({ instruction: "nihao", activeBookId: "demo-book", sessionId: "agent-session-1" }),
-    });
-
-    expect(response.status).toBe(502);
-    await expect(response.json()).resolves.toEqual({
-      error: {
-        code: "AGENT_LLM_ERROR",
-        message: upstreamError,
-      },
-      response: upstreamError,
-    });
-    expect(chatCompletionMock).not.toHaveBeenCalled();
-  });
-
-  it("returns malformed Gemini function-call errors without replacing them with an empty-response probe", async () => {
-    const upstreamError = "Provider finish_reason: function_call_filter: MALFORMED_FUNCTION_CALL";
-    runAgentSessionMock.mockResolvedValueOnce({
-      responseText: "",
-      errorMessage: upstreamError,
-      messages: [{ role: "assistant", content: [], stopReason: "error", errorMessage: upstreamError }],
-    });
-
-    const { createStudioServer } = await import("./server.js");
-    const app = createStudioServer(cloneProjectConfig() as never, root);
-
-    const response = await app.request("http://localhost/api/v1/agent", {
-      method: "POST",
-      headers: { "Content-Type": "application/json" },
-      body: JSON.stringify({ instruction: "nihao", activeBookId: "demo-book", sessionId: "agent-session-1" }),
-    });
-
-    expect(response.status).toBe(502);
-    await expect(response.json()).resolves.toEqual({
-      error: {
-        code: "AGENT_LLM_ERROR",
-        message: upstreamError,
-      },
-      response: upstreamError,
-    });
-    expect(chatCompletionMock).not.toHaveBeenCalled();
-  });
-
-  it("falls back to plain chat when the tool-agent returns empty text", async () => {
-    runAgentSessionMock.mockResolvedValueOnce({
-      responseText: "",
-      messages: [{ role: "user", content: "nihao" }],
-    });
-    chatCompletionMock.mockResolvedValueOnce({
-      content: "你好！",
-      usage: { promptTokens: 1, completionTokens: 1, totalTokens: 2 },
-    });
-
-    const { createStudioServer } = await import("./server.js");
-    const app = createStudioServer(cloneProjectConfig() as never, root);
-
-    const response = await app.request("http://localhost/api/v1/agent", {
-      method: "POST",
-      headers: { "Content-Type": "application/json" },
-      body: JSON.stringify({ instruction: "nihao", activeBookId: "demo-book", sessionId: "agent-session-1" }),
-    });
-
-    expect(response.status).toBe(200);
-    await expect(response.json()).resolves.toEqual({
-      response: "你好！",
-      session: { sessionId: "agent-session-1" },
-    });
-  });
-
-  it("migrates and exposes a book created by architect even when the final agent text is empty", async () => {
-    const orphanSession = {
-      sessionId: "agent-session-1",
-      bookId: null,
-      title: null,
-      messages: [],
-      events: [],
-      draftRounds: [],
-      createdAt: 1,
-      updatedAt: 1,
-    };
-    loadBookSessionMock.mockResolvedValue(orphanSession);
-    appendBookSessionMessageMock.mockImplementation((session: unknown) => session);
-    migrateBookSessionMock.mockResolvedValue({
-      ...orphanSession,
-      bookId: "new-book",
-    });
-    loadBookConfigMock.mockImplementation(async (bookId?: string) => ({
-      id: bookId ?? "new-book",
-      title: "New Book",
-      platform: "qidian",
-      genre: "urban",
-      status: "outlining",
-      targetChapters: 100,
-      chapterWordCount: 3000,
-      createdAt: "2026-04-12T00:00:00.000Z",
-      updatedAt: "2026-04-12T00:00:00.000Z",
-    }));
-    runAgentSessionMock.mockImplementationOnce(async (config: { onEvent?: (event: unknown) => void }) => {
-      config.onEvent?.({
-        type: "tool_execution_start",
-        toolCallId: "tool-1",
-        toolName: "sub_agent",
-        args: { agent: "architect", title: "New Book" },
-      });
-      config.onEvent?.({
-        type: "tool_execution_end",
-        toolCallId: "tool-1",
-        toolName: "sub_agent",
-        isError: false,
-        result: {
-          content: [{ type: "text", text: "Book created." }],
-          details: { kind: "book_created", bookId: "new-book", title: "New Book" },
-        },
-      });
-      return {
-        responseText: "",
-        messages: [{ role: "user", content: "/new New Book" }],
-      };
-    });
-    chatCompletionMock.mockResolvedValueOnce({
-      content: "建书完成。",
-      usage: { promptTokens: 1, completionTokens: 1, totalTokens: 2 },
-    });
-
-    const { createStudioServer } = await import("./server.js");
-    const app = createStudioServer(cloneProjectConfig() as never, root);
-
-    const response = await app.request("http://localhost/api/v1/agent", {
-      method: "POST",
-      headers: { "Content-Type": "application/json" },
-      body: JSON.stringify({ instruction: "写一本都市商战", sessionId: "agent-session-1" }),
-    });
-
-    expect(response.status).toBe(200);
-    expect(migrateBookSessionMock).toHaveBeenCalledWith(root, "agent-session-1", "new-book");
-    await expect(response.json()).resolves.toMatchObject({
-      response: "建书完成。",
-      session: {
-        sessionId: "agent-session-1",
-        activeBookId: "new-book",
-      },
-    });
-  });
-
-  it("rejects /api/v1/agent requests without sessionId", async () => {
-    const { createStudioServer } = await import("./server.js");
-    const app = createStudioServer(cloneProjectConfig() as never, root);
-
-    const response = await app.request("http://localhost/api/v1/agent", {
-      method: "POST",
-      headers: { "Content-Type": "application/json" },
-      body: JSON.stringify({ instruction: "continue", activeBookId: "demo-book" }),
-    });
-
-    expect(response.status).toBe(400);
-    await expect(response.json()).resolves.toEqual({
-      error: {
-        code: "SESSION_ID_REQUIRED",
-        message: "sessionId is required",
-      },
-    });
-  });
-
-  it("returns the shared interaction session state", async () => {
-    loadProjectSessionMock.mockResolvedValue({
-      sessionId: "session-2",
-      projectRoot: root,
-      activeBookId: "demo-book",
-      automationMode: "auto",
-      messages: [
-        { role: "user", content: "continue", timestamp: 1 },
-      ],
-    });
-    resolveSessionActiveBookMock.mockResolvedValue("demo-book");
-
-    const { createStudioServer } = await import("./server.js");
-    const app = createStudioServer(cloneProjectConfig() as never, root);
-
-    const response = await app.request("http://localhost/api/v1/interaction/session");
-
-    expect(response.status).toBe(200);
-    await expect(response.json()).resolves.toMatchObject({
-      session: expect.objectContaining({
-        activeBookId: "demo-book",
-        automationMode: "auto",
-      }),
-      activeBookId: "demo-book",
-    });
-  });
-
-  it("returns creation-draft state through the shared interaction session endpoint", async () => {
-    loadProjectSessionMock.mockResolvedValue({
-      sessionId: "session-3",
-      projectRoot: root,
-      automationMode: "semi",
-      creationDraft: {
-        concept: "港风商战悬疑，主角从灰产洗白。",
-        title: "夜港账本",
-        nextQuestion: "你更想写长篇连载，还是十来章能收住？",
-        missingFields: ["targetChapters"],
-        readyToCreate: false,
-      },
-      messages: [],
-    });
-    resolveSessionActiveBookMock.mockResolvedValue(undefined);
-
-    const { createStudioServer } = await import("./server.js");
-    const app = createStudioServer(cloneProjectConfig() as never, root);
-
-    const response = await app.request("http://localhost/api/v1/interaction/session");
-
-    expect(response.status).toBe(200);
-    await expect(response.json()).resolves.toMatchObject({
-      session: expect.objectContaining({
-        creationDraft: expect.objectContaining({
-          title: "夜港账本",
-          nextQuestion: "你更想写长篇连载，还是十来章能收住？",
-        }),
-      }),
     });
   });
 });
